add: spawn fish button & money resource logic

+++ Assets/Resources/Entities/Battlers/BattlersStat.cs
@@ -13,7 +13,7 @@ public class BattlersStat : ScriptableObject
         return moveSpeed;
     }
 
-    public float GetPrice()
+    public int GetPrice()
     {
         return price;
     }
